+++ src/backend/mod.rs
@@ -1,7 +1,7 @@
 mod cassandra;
 mod dynamodb;
 mod mysql;
-mod postgres;
+mod postgresql;
 mod redis;
 mod tidb;
 mod tikv;
@@ -9,7 +9,7 @@ mod tikv;
 pub use cassandra::Cassandra;
 pub use dynamodb::DynamoDB;
 pub use mysql::MySQL;
-pub use postgres::PostgreSQL;
+pub use postgresql::PostgreSQL;
 pub use redis::Redis;
 pub use tidb::TiDB;
 pub use tikv::TiKV;
